refactor: 품목 코드 데이터 타입 변경 #17

+++ backend/smart-logistics/src/main/java/com/douzon/smartlogistics/domain/entity/Item.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class Item {
 
-    private final Long itemCode;
+    private final Integer itemCode;
     private final String itemName;
     private final String spec;
     private final String unit;
